Refactor and bug fixed in 'findocc1()' and 'sfind1()'

+++ biostr.h
@@ -23,6 +23,6 @@ unsigned sfind(char *tosearch, char *tofind, unsigned startpos);
 unsigned repfind(char *tosearch, char tofind, unsigned minlen, unsigned startpos, STAT_T *actlen);
 
 
-unsigned long sfind1(char *targ, SRCH_T *pat, unsigned long start);
+unsigned sfind1(char *targ, SRCH_T *pat, unsigned long start);
 
 #endif /* biostr_h */
